+++ kernel-rs/src/plic.rs
@@ -11,8 +11,8 @@ use crate::{
 /// the riscv Platform Level Interrupt Controller (PLIC).
 pub unsafe fn plicinit() {
     // set desired IRQ priorities non-zero (otherwise disabled).
-    *((PLIC + (UART0_IRQ * 4) as i64) as *mut u32) = 1 as i32 as u32;
-    *((PLIC + (VIRTIO0_IRQ * 4) as i64) as *mut u32) = 1 as i32 as u32;
+    *((PLIC + (UART0_IRQ * 4) as i64) as *mut u32) = 1 as u32;
+    *((PLIC + (VIRTIO0_IRQ * 4) as i64) as *mut u32) = 1 as u32;
 }
 
 pub unsafe fn plicinithart() {
@@ -20,10 +20,10 @@ pub unsafe fn plicinithart() {
 
     // set uart's enable bit for this hart's S-mode.
     *(plic_senable(hart) as *mut u32) =
-        ((1 as i32) << UART0_IRQ | (1 as i32) << VIRTIO0_IRQ) as u32;
+        ((1) << UART0_IRQ | (1) << VIRTIO0_IRQ) as u32;
 
     // set this hart's S-mode priority threshold to 0.
-    *(plic_spriority(hart) as *mut u32) = 0 as i32 as u32;
+    *(plic_spriority(hart) as *mut u32) = 0 as u32;
 }
 
 /// return a bitmap of which IRQs are waiting
